강의 #8까지

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 
 export const home = async (req, res) => {
@@ -9,7 +10,8 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
     const {id} = req.params; //ES6 문법
     // const id = req.params.id; 위와 동일
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
@@ -18,17 +20,25 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("videos/edit", {pageTitle: `Edit ⇢ ${video.title}`, video});
 };
 export const postEdit = async (req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const {title, description, hashtags} = req.body;
     if (!await Video.exists({_id: id})) {
         return res.render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, description, hashtags: Video.formatHashtags(hashtags)
@@ -40,13 +50,22 @@ export const getUpload = (req, res) => {
     return res.render("videos/upload", {pageTitle: "Upload Video"});
 }
 export const postUpload = async (req, res) => {
-    const {title, description, hashtags} = req.body;
+    const {
+        session: {user: {_id}},
+        file: {path: fileUrl},
+        body: {title, description, hashtags}
+    } = req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("videos/upload", {pageTitle: "Upload Video", errorMessage: error._message});
@@ -55,8 +74,15 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(400).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
-    // delete video
     return res.redirect("/");
 }
